Frontend Service - Update: Fixed authentication logic and landing page UI.
