+++ Frontend/components/UI/LandingPage/Footer.vue
@@ -2,12 +2,9 @@
     <div class="bg-white rounded-lg shadow m-4">
         <div class="w-full mx-auto max-w-screen-xl p-4 md:flex md:items-center md:justify-between">
             <span class="text-sm text-gray-500 sm:text-center"
-                >© 2023 <a href="https://flowbite.com/" class="hover:underline">Flowbite™</a>. All
-                Rights Reserved.
+                >© 2024 <a href="#" class="hover:underline">IoTunnel</a>. All Rights Reserved.
             </span>
-            <ul
-                class="flex flex-wrap items-center mt-3 text-sm font-medium text-gray-500 sm:mt-0"
-            >
+            <ul class="flex flex-wrap items-center mt-3 text-sm font-medium text-gray-500 sm:mt-0">
                 <li>
                     <a href="#" class="hover:underline me-4 md:me-6">About</a>
                 </li>
