+++ Frontend/composables/Encryption/index.ts
@@ -3,15 +3,31 @@ import { useRuntimeConfig } from "#imports";
 
 function encrypt(text: string) {
     const config = useRuntimeConfig();
-    const passphrase = config.public.privateKey + config.public.publicKey;
-    return CryptoJS.AES.encrypt(text, passphrase).toString();
+    const nonce = CryptoJS.lib.WordArray.random(16).toString();
+    const passphrase = CryptoJS.SHA256(
+        config.public.privateKey + config.public.publicKey + nonce
+    ).toString();
+    const encrypted = CryptoJS.AES.encrypt(text, passphrase, { nonce: nonce });
+    return encrypted.toString() + ":" + nonce;
 }
 
 function decrypt(ciphertext: string) {
     const config = useRuntimeConfig();
-    const passphrase = config.public.privateKey + config.public.publicKey;
-    const bytes = CryptoJS.AES.decrypt(ciphertext, passphrase);
-    return bytes.toString(CryptoJS.enc.Utf8);
+    const ciphertextParts = ciphertext.split(":");
+    if (ciphertextParts.length !== 2) {
+        return "";
+    }
+    const encrypted = ciphertextParts[0];
+    const nonce = ciphertextParts[1];
+    const passphrase = CryptoJS.SHA256(
+        config.public.privateKey + config.public.publicKey + nonce
+    ).toString();
+    try {
+        const decrypted = CryptoJS.AES.decrypt(encrypted, passphrase, { nonce: nonce });
+        return decrypted.toString(CryptoJS.enc.Utf8);
+    } catch (error) {
+        return "";
+    }
 }
 
 export { encrypt, decrypt };
