+++ Frontend/pages/admins/auth/signin.vue
@@ -1,5 +1,5 @@
 <script setup>
-import SignInAdmin from "~/components/Authentication/SignInAdmin.vue";
+import SignInAdmin from "~/components/UI/Authentication/SignInAdmin.vue";
 </script>
 
 <template>
