+++ Frontend/pages/admins/dashboard.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+import AdminAuth from "~/components/Core/Authentication/AdminAuth.vue";
+</script>
+
+<template>
+    <AdminAuth />
+    <h1>Dashboard</h1>
+</template>
