+++ Frontend/pages/users/dashboard.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+import UserAuth from "~/components/Core/Authentication/UserAuth.vue";
+</script>
+
+<template>
+    <UserAuth />
+    <h1>Dashboard</h1>
+</template>
